requirements.txt
openai
streamlit
pandas
python-dotenv
azure-core
azure-search
azure-search-documents